feat: Repackage as a Sphinx extension
Repackage the project as a Sphinx extension so that it can be consumed
by python package managers easily without the nead for git submodules.

This changes the build process slightly for this theme, so see the
changes to build_dist.sh as well as the updated README.md and demo.